optimize preinstall check
